+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [1.0.2] - 2026-01-29
+
+### Summary
+
+test(tests): update pyproject.toml, test_cli_options.py
+
+### Changes
+
+- update pyproject.toml
+- update tests/test_cli_options.py
+
 ## [1.0.1] - 2026-01-29
 
 - docs: update README